Media Queries, Loaders, Error Handling, Mobile Nav
*Media Queries for Home, Login, and Register.

*Error handling
Basic error handling page that shows status and a (not so) dynamic

*Router Loaders
Set up with a provisory API. Just to build the SPA structure of the project.
Example: Dynamic route to access a user's profile page, which has a loader to fetch that user's information.

*Mobile Navigation
Burger menu built (not finished)